Cleanup

+++ app/controllers/vehicles_controller.rb
@@ -1,9 +1,9 @@
 class VehiclesController < ApplicationController
-  layout 'admin_material'
-
   before_action :signed_in_user
 
   def index
+    authorize Vehicle, :index?
+
     @vehicles = policy_scope(Vehicle).includes(:expirations)
 
     respond_to do |format|
@@ -13,11 +13,15 @@ class VehiclesController < ApplicationController
   end
 
   def show
+    authorize Vehicle, :show?
+
     @vehicle = policy_scope(Vehicle).find(params[:id])
     @equipment_requests = @vehicle.equipment_requests
   end
 
   def create
+    authorize Vehicle, :create?
+
     vehicle = Vehicle.create(vehicle_params.merge({ organization_id: OrganizationContext.current_organization.id }))
 
     render json: vehicle
@@ -25,6 +29,8 @@ class VehiclesController < ApplicationController
 
 
   def update
+    authorize Vehicle, :update?
+
     vehicle = policy_scope(Vehicle).find(params[:id])
     vehicle.update(vehicle_params)
 
@@ -32,6 +38,8 @@ class VehiclesController < ApplicationController
   end
 
   def destroy
+    authorize Vehicle, :destroy?
+
     vehicle = Vehicle.find(params[:id])
     vehicle.destroy
